+++ packages/fastify-generators/src/generators/core/fastify-sentry/index.ts
@@ -73,10 +73,15 @@ const FastifySentryGenerator = createGeneratorWithTasks({
     taskBuilder.addTask({
       name: 'server',
       dependencies: {
+        node: nodeProvider,
         fastifyServer: fastifyServerProvider,
         errorHandlerServiceSetup: errorHandlerServiceSetupProvider,
       },
-      run({ errorHandlerServiceSetup, fastifyServer }) {
+      run({ node, errorHandlerServiceSetup, fastifyServer }) {
+        if (!node.isEsm()) {
+          fastifyServer.addInitializerBlock("import './instrument.js';\n");
+        }
+
         fastifyServer.addPrePluginBlock(
           TypescriptCodeUtils.createBlock(
             `Sentry.setupFastifyErrorHandler(fastify);
@@ -129,14 +134,14 @@ const FastifySentryGenerator = createGeneratorWithTasks({
         const shouldLogToSentryBlocks: TypescriptCodeBlock[] = [];
 
         node.addPackages({
-          '@sentry/core': '8.34.0',
-          '@sentry/node': '8.34.0',
-          '@sentry/profiling-node': '8.34.0',
+          '@sentry/core': '8.41.0',
+          '@sentry/node': '8.41.0',
+          '@sentry/profiling-node': '8.41.0',
           lodash: '4.17.21',
         });
 
         node.addDevPackages({
-          '@sentry/types': '8.34.0',
+          '@sentry/types': '8.41.0',
           '@types/lodash': '4.17.7',
         });