+++ packages/project-builder-lib/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @halfdomelabs/project-builder-lib
 
+## 0.8.13
+
+### Patch Changes
+
+- 6ca94da: Upgrade Zod to 3.24.1
+- 98518dd: Upgrade nanoid to 5.0.9
+- Updated dependencies [6ca94da]
+- Updated dependencies [354f4c9]
+  - @halfdomelabs/ui-components@0.5.1
+  - @halfdomelabs/sync@0.7.12
+
+## 0.8.12
+
+### Patch Changes
+
+- 77d9399: Upgrade ESLint to v9 and use updated Linter configurations
+- Updated dependencies [77d9399]
+- Updated dependencies [f11b044]
+- Updated dependencies [dc74c47]
+- Updated dependencies [05fbd9c]
+  - @halfdomelabs/ui-components@0.5.0
+  - @halfdomelabs/sync@0.7.11
+
 ## 0.8.11
 
 ### Patch Changes