+++ packages/sync/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @halfdomelabs/sync
 
+## 0.7.12
+
+### Patch Changes
+
+- 6ca94da: Upgrade Zod to 3.24.1
+- 354f4c9: [internal] Refactor formatter provider to become part of generator output
+
+## 0.7.11
+
+### Patch Changes
+
+- 77d9399: Upgrade ESLint to v9 and use updated Linter configurations
+
 ## 0.7.10
 
 ### Patch Changes